fix map size & refactoring extension
