+++ Plogging/GPS/GPS.storyboard
@@ -38,7 +38,7 @@
                                     <action selector="onTouch:" destination="5vR-lz-Sff" eventType="touchUpInside" id="gWe-ou-yGy"/>
                                 </connections>
                             </button>
-                            <mapView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" fixedFrame="YES" mapType="standard" translatesAutoresizingMaskIntoConstraints="NO" id="GLX-Le-nBw">
+                            <mapView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" fixedFrame="YES" mapType="standard" showsUserLocation="YES" showsScale="YES" translatesAutoresizingMaskIntoConstraints="NO" id="GLX-Le-nBw">
                                 <rect key="frame" x="67" y="456" width="280" height="340"/>
                                 <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                             </mapView>
