--- Plogging/GPS/MapkitDelegate.swift
@@ -1,37 +1,0 @@
-//
-// Created by KHU_TAEWOO on 2021/01/11.
-//
-
-import Foundation
-import CoreLocation
-import MapKit
-
-extension MapkitViewController: CLLocationManagerDelegate, MKMapViewDelegate {
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        for newLocation in locations {
-            let howRecent = newLocation.timestamp.timeIntervalSinceNow
-            guard newLocation.horizontalAccuracy < 20 && abs(howRecent) < 10 else { continue }
-
-            if let lastLocation = locationList.last {
-                let delta = newLocation.distance(from: lastLocation)
-                distance = distance + Measurement(value: delta, unit: UnitLength.meters)
-            }
-
-            locationList.append(newLocation)
-            
-        }
-    }
-
-    public func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
-        guard let polyline = overlay as? MKPolyline else {
-            return MKOverlayRenderer(overlay: overlay)
-        }
-        let renderer = MKPolylineRenderer(polyline: polyline)
-        renderer.strokeColor = .red
-        renderer.lineWidth = 3
-        return renderer
-    }
-
-}
-
